requirements.txt
python==3.10
ollama==0.2.0
phidata==2.4.14
streamlit==1.35.0